rename actiontypes files
