+++ src/redux/reducers/incomeReducer.js
@@ -1,4 +1,4 @@
-import * as IncomeTypes from "../actiontypes/incomes";
+import * as IncomeTypes from "../actiontypes/incomesActionTypes";
 
 const initialState = {
   incomes: [],
