+++ src/redux/sagas/incomeSaga.js
@@ -1,6 +1,6 @@
 import { takeEvery, put, call } from "redux-saga/effects";
 import { getNonDeletedIncomes, saveIncome, deleteIncome } from "../../api/incomeApi";
-import * as IncomeActionTypes from "../actiontypes/incomes";
+import * as IncomeActionTypes from "../actiontypes/incomesActionTypes";
 
 function* loadUserIncomes() {
   const incomes = yield call(getNonDeletedIncomes);
@@ -39,7 +39,6 @@ function* deleteIncomeOnApi(income) {
 }
 
 export function* watchLoadIncomes() {
-  // Each time someone calls LOAD_USER, will perfom getUserFromFirebase
   yield takeEvery(IncomeActionTypes.LOAD_INCOMES, loadUserIncomes);
 }
 export function* watchSaveIncome() {
